added test for copy
