+++ Oras/Copy.cs
@@ -1,9 +1,9 @@
-﻿using static Oras.Content.Content;
-using Oras.Interfaces;
+﻿using Oras.Interfaces;
 using Oras.Models;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using static Oras.Content.Content;
 
 namespace Oras
 {
