+++ Oras/Memory/MemoryGraph.cs
@@ -1,11 +1,11 @@
-﻿using static Oras.Content.Content;
-using Oras.Interfaces;
+﻿using Oras.Interfaces;
 using Oras.Models;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using static Oras.Content.Content;
 
 namespace Oras.Memory
 {
